fix(raycast): split dda(), move tiles_len(), norm fixes
